// src/api/stetsom/model/patchApiTechnicalAssistancesIdBody.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */

export type PatchApiTechnicalAssistancesIdBody = {
  /**
   * @minLength 1
   * @maxLength 200
   */
  name?: string;
  /**
   * @minLength 1
   * @maxLength 500
   */
  address?: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  city?: string;
  /**
   * @minLength 2
   * @maxLength 2
   */
  state?: string;
  /**
   * @minLength 8
   * @maxLength 9
   */
  zip?: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /**
   * @minLength 1
   * @maxLength 20
   */
  phone?: string;
  /**
   * @maxLength 20
   * @nullable
   */
  phone2?: string | null;
  /**
   * @nullable
   * @pattern ^(?!\.)(?!.*\.\.)([A-Za-z0-9_'+\-\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
   */
  email?: string | null;
  /**
   * @maxLength 200
   * @nullable
   */
  specialty?: string | null;
  is_active?: boolean;
};