// src/api/stetsom/model/technicalAssistance.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */

export interface TechnicalAssistance {
  id: string;
  name: string;
  address: string;
  city: string;
  state: string;
  zip: string;
  lat?: number;
  lng?: number;
  phone: string;
  phone2?: string;
  email?: string;
  specialty?: string;
  is_active: boolean;
  created_at: string;
  updated_at: string;
}